requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
aiohttp
python-multipart
